+++ src/main/java/net/sf/jabref/gui/entryeditor/FieldExtraComponents.java
@@ -131,14 +131,17 @@ public class FieldExtraComponents {
         JTextComponent url = (JTextComponent) fieldEditor;
 
         DocumentListener documentListener = new DocumentListener() {
+            @Override
             public void changedUpdate(DocumentEvent documentEvent) {
                 checkUrl();
             }
 
+            @Override
             public void insertUpdate(DocumentEvent documentEvent) {
                 checkUrl();
             }
 
+            @Override
             public void removeUpdate(DocumentEvent documentEvent) {
                 checkUrl();
             }
@@ -200,14 +203,17 @@ public class FieldExtraComponents {
         JTextComponent doi = (JTextComponent) fieldEditor;
 
         DocumentListener documentListener = new DocumentListener() {
+            @Override
             public void changedUpdate(DocumentEvent documentEvent) {
                 checkDoi();
             }
 
+            @Override
             public void insertUpdate(DocumentEvent documentEvent) {
                 checkDoi();
             }
 
+            @Override
             public void removeUpdate(DocumentEvent documentEvent) {
                 checkDoi();
             }
@@ -362,6 +368,88 @@ public class FieldExtraComponents {
     }
 
     /**
+     * Return a dropdown list with the alternatives for editor type fields
+     *
+     * @param fieldEditor
+     * @param entryEditor
+     * @return
+     */
+
+    public static Optional<JComponent> getEditorTypeExtraComponent(FieldEditor fieldEditor, EntryEditor entryEditor) {
+        final String[] optionValues = {"", "editor", "compiler", "founder", "continuator", "redactor", "reviser",
+                "collaborator"};
+        final String[] optionDescriptions = {Localization.lang("Select"), Localization.lang("Editor"),
+                Localization.lang("Compiler"), Localization.lang("Founder"), Localization.lang("Continuator"),
+                Localization.lang("Redactor"), Localization.lang("Reviser"), Localization.lang("Collaborator")};
+        JComboBox<String> editorType = new JComboBox<>(optionDescriptions);
+        editorType.addActionListener(actionEvent -> {
+            fieldEditor.setText(optionValues[editorType.getSelectedIndex()]);
+            entryEditor.updateField(fieldEditor);
+        });
+        return Optional.of(editorType);
+
+    }
+
+    /**
+     * Return a dropdown list with the alternatives for pagination type fields
+     *
+     * @param fieldEditor
+     * @param entryEditor
+     * @return
+     */
+
+    public static Optional<JComponent> getPaginationExtraComponent(FieldEditor fieldEditor, EntryEditor entryEditor) {
+        final String[] optionValues = {"", "page", "column", "line", "verse", "section", "paragraph", "none"};
+        final String[] optionDescriptions = {Localization.lang("Select"), Localization.lang("Page"),
+                Localization.lang("Column"), Localization.lang("Line"), Localization.lang("Verse"),
+                Localization.lang("Section"), Localization.lang("Paragraph"), Localization.lang("None")};
+        JComboBox<String> pagination = new JComboBox<>(optionDescriptions);
+        pagination.addActionListener(actionEvent -> {
+            fieldEditor.setText(optionValues[pagination.getSelectedIndex()]);
+            entryEditor.updateField(fieldEditor);
+        });
+        return Optional.of(pagination);
+    }
+
+    /**
+     * Return a dropdown list with the alternatives for pagination type fields
+     *
+     * @param fieldEditor
+     * @param entryEditor
+     * @return
+     */
+
+    public static Optional<JComponent> getTypeExtraComponent(FieldEditor fieldEditor, EntryEditor entryEditor,
+            boolean isPatent) {
+        String[] optionValues;
+        String[] optionDescriptions;
+        if (isPatent) {
+            optionValues = new String[] {"", "patent", "patentde", "patenteu", "patentfr", "patentuk", "patentus",
+                    "patreq", "patreqde", "patreqeu", "patreqfr", "patrequk", "patrequs"};
+            optionDescriptions = new String[] {Localization.lang("Select"), Localization.lang("Patent"),
+                    Localization.lang("German patent"), Localization.lang("European patent"),
+                    Localization.lang("French patent"), Localization.lang("British patent"),
+                    Localization.lang("U.S. patent"), Localization.lang("Patent request"),
+                    Localization.lang("German patent request"), Localization.lang("European patent request"),
+                    Localization.lang("French patent request"), Localization.lang("British patent request"),
+                    Localization.lang("U.S. patent request")};
+        } else {
+            optionValues = new String[] {"", "mathesis", "phdthesis", "candthesis", "techreport", "resreport",
+                    "software", "datacd", "audiocd"};
+            optionDescriptions = new String[] {Localization.lang("Select"), Localization.lang("Master's thesis"),
+                    Localization.lang("PhD thesis"), Localization.lang("Candidate thesis"),
+                    Localization.lang("Technical report"), Localization.lang("Research report"),
+                    Localization.lang("Software"), Localization.lang("Data CD"), Localization.lang("Audio CD")};
+        }
+        JComboBox<String> type = new JComboBox<>(optionDescriptions);
+        type.addActionListener(actionEvent -> {
+            fieldEditor.setText(optionValues[type.getSelectedIndex()]);
+            entryEditor.updateField(fieldEditor);
+        });
+        return Optional.of(type);
+    }
+
+    /**
      * Return a dropdown list with the gender alternatives for fields with GENDER
      *
      * @param fieldEditor
@@ -371,7 +459,8 @@ public class FieldExtraComponents {
 
     public static Optional<JComponent> getGenderExtraComponent(FieldEditor fieldEditor, EntryEditor entryEditor) {
         final String[] optionValues = {"", "sf", "sm", "sp", "pf", "pm", "pn", "pp"};
-        final String[] optionDescriptions = {"", Localization.lang("Female name"), Localization.lang("Male name"),
+        final String[] optionDescriptions = {Localization.lang("Select"), Localization.lang("Female name"),
+                Localization.lang("Male name"),
                 Localization.lang("Neuter name"), Localization.lang("Female names"), Localization.lang("Male names"),
                 Localization.lang("Neuter names"), Localization.lang("Mixed names")};
         JComboBox<String> gender = new JComboBox<>(optionDescriptions);